🏇✨ search 결과 나오는 컴포넌트 추가 router설정 할 예정
