--- src/NASAapi.js
@@ -1,4 +1,0 @@
-
-export const Api = function  (query){
-    return `https://api.nasa.gov/${query}/apod?api_key=iT6Yq1QPqgLXhA3PySqO7ZkGvQzSezLsnfJ0zUkt`
-}
